requirements.txt
flask==3.0.2
requests==2.31.0
pandas==2.2.1
yfinance==0.2.37
polygon-api-client==1.12.4
python-dotenv==1.0.1
supabase==2.3.0
flask-cors==4.0.0 